+++ tests/tests_fabric/plugins/precision/test_deepspeed.py
@@ -23,11 +23,11 @@ from lightning.fabric.utilities.types import Steppable
 
 def test_invalid_precision_with_deepspeed_precision():
     with pytest.raises(ValueError, match="is not supported in DeepSpeed. `precision` must be one of"):
-        DeepSpeedPrecision(precision=64)
+        DeepSpeedPrecision(precision="64-true")
 
 
 def test_deepspeed_precision_backward():
-    precision = DeepSpeedPrecision(precision=32)
+    precision = DeepSpeedPrecision(precision="32-true")
     tensor = Mock()
     model = Mock()
     precision.backward(tensor, model, "positional-arg", keyword="arg")
@@ -45,7 +45,7 @@ def test_deepspeed_engine_is_steppable(engine):
 
 
 def test_deepspeed_precision_optimizer_step():
-    precision = DeepSpeedPrecision(precision=32)
+    precision = DeepSpeedPrecision(precision="32-true")
     optimizer = model = Mock()
     precision.optimizer_step(optimizer, lr_kwargs=dict())
     model.step.assert_called_once_with(lr_kwargs=dict())